Pass Column down to the generators
